+++ R/elo.colley.R
@@ -81,8 +81,8 @@ elo.colley <- function(formula, data, family = "binomial", weights, na.action, s
 
     for(i in setdiff(seq_len(max(grp2)), seq_len(skip)))
     {
-      if(i == 1) next
-      sbst <- grp2 %in% 1:(i-1)
+      if(i == 0) next
+      sbst <- grp2 %in% 0:(i-1)
       dat.tmp <- dat
       dat.tmp$winsA <- dat.tmp$winsA[sbst]
       dat.tmp$k <- dat.tmp$k[sbst, , drop = FALSE]
@@ -103,6 +103,7 @@ elo.colley <- function(formula, data, family = "binomial", weights, na.action, s
       ftd[grp2 == i] <- apply(cbind(difference, adj)[grp2 == i, , drop=FALSE], 1, function(x) sum(x * coeff, na.rm = TRUE))
     }
     out$running.values <- mc.glm$family$linkinv(ftd)
+    attr(out$running.values, "group") <- grp2
   }
 
   structure(out, class = c(if(running) "elo.running", "elo.colley"))