+++ R/internal.functions.R
@@ -177,7 +177,8 @@ group_to_int <- function(grp, skip)
   grp2 <- rev(cumsum(rev(grp2)))
   mx <- max(grp2)
   if(skip > mx || skip < 0) stop("skip must be between 0 and ", mx, " (inclusive)")
-  mx + 1 - grp2 # from mx : 1 to 1 : mx
+  out <- mx + 1 - grp2 # from mx : 1 to 1 : mx
+  replace(out, out <= skip,)
 }